backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
tensorflow==2.16.1
pillow==10.3.0
pillow-avif-plugin==1.4.3
numpy==1.26.4
gdown==5.1.0